siloenced some compiler warnings

+++ cmscoffeeaccountmodel.cpp
@@ -29,7 +29,7 @@ void CMSCoffeeAccountModel::sortByDate()
     std::sort(entries_.begin(), entries_.end(),
               [&](const CMSCoffeeAccountEntry* a, const CMSCoffeeAccountEntry* b)
     {
-        int days = a->getDate().daysTo(b->getDate());
+        long long days = a->getDate().daysTo(b->getDate());
         if (days!=0) return (days<0);
 
         CMSCoffeeUser * ua = userModel_->userByUUID(a->getUUID());
